Update news item image handling and form submission
The changes involve handling of images within news items and form submission process. The ability to convert image to base64 and display it in the news feed has been added. The news creation form has been modified to support multipart/form-data and removed the ajax call on submission. Improved labels for the fields in the form to be more informative.