Add ImageService for image downloading and saving
A new service, ImageService, has been added to the POSTerminal Services. This service is responsible for downloading images from a given URL and saving them into the database. It implements the IImageService interface which includes the `DownloadAndSaveImage` method. Exception handling is also put in place to manage failures during the image fetching process.