Add ImageId column to NewsItems table
This commit introduces a new nullable ImageId column to the NewsItems table in the POSTerminal application, in order to keep a reference to related images. It also includes the creation of the ImageItems table to store all images, along with the necessary migrational changes. These updates will improve the management and retrieval of image data associated with NewsItems.