Implement Markdown rendering and enhance forms in Jobs views
Additional features have been included in the Jobs views: the parsing and rendering of text in a Markdown format and the enhancement of the forms for a better user experience. The enhancement of forms consists of improving the input types for better data handling, and increasing user feedback accuracy. Further additions include the integration of the Westwind.AspNetCore.Markdown library for better Markdown rendering.