Integrate image uploading to JobsController
The code has been updated to allow image uploads in the JobsController. The 'Create' function now includes functionality to handle IFormFile images. If no file is provided but there is a URL present in 'PicSrc', the image is downloaded and saved using the ImageService. If an ImageModel was generated from either of these processes, it is attached to the JobsItem before being saved in the context.